code restructuring moved Join and Union functions to teh core package making them private
