+++ dstream-api/src/main/java/dstream/StreamUnionFunction.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dstream.function;
+package dstream;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -28,7 +28,7 @@ import dstream.utils.Tuples.Tuple2;
  * Implementation of {@link SerFunction} which will union multiple streams
  * while applying user functionality at check points (see this{@link #addCheckPoint(int)}.
  */
-public class StreamUnionFunction extends AbstractMultiStreamProcessingFunction {
+class StreamUnionFunction extends AbstractMultiStreamProcessingFunction {
 	private static final long serialVersionUID = -2955908820407886806L;
 	
 	private final boolean distinct;
